+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/index/rondb/IndexRecordKey.java
@@ -34,7 +34,7 @@ import java.util.Collection;
 
 @Entity
 @Table(name = "index_record_key",
-        indexes = @Index(name = "record_key_index", columnList = "key"))
+        indexes = @Index(columnList = "record_key", unique = false))
 @NamedQueries({
         @NamedQuery(name = "RecordKey.getByKey",
                 query = "SELECT recordKey FROM IndexRecordKey recordKey WHERE recordKey.key = :key"),
@@ -47,7 +47,7 @@ public class IndexRecordKey implements Serializable {
   @Column(name = "id")
   private Long id;
 
-  @Column(name = "record_key", columnDefinition = "VARBINARY(255) NOT NULL")
+  @Column(name = "record_key", columnDefinition = "VARBINARY(255)", nullable = false)
   private byte[] key;
 
   @OneToMany(cascade = CascadeType.ALL, mappedBy = "recordKey")
