+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/index/rondb/IndexRecordPartition.java
@@ -34,7 +34,7 @@ import java.util.Collection;
 
 @Entity
 @Table(name = "index_record_partition",
-        indexes = @Index(name = "record_partition_index", columnList = "path"))
+        indexes = @Index(columnList = "path", unique = false))
 @NamedQueries({
         @NamedQuery(name = "RecordPartition.getByPath",
                 query = "SELECT recordPartition FROM IndexRecordPartition recordPartition WHERE recordPartition.path = :path")})
